Design a Map Unit

This function calls similar_soils() iteratively to estimate the proportions of observations corresponding to distinct groups of similar soils. Use it exhaustively group all observations (for example, within a map unit or other conceptual unit of interest)

Usage

design_mapunit(
  x,
  mapping,
  idname = "id",
  component_labels = SOILmilaR::greekletters[[1]],
  ...
)

Arguments

x

A data.frame or a SoilProfileCollection.

mapping

A named list of rating functions.

idname

The name of the ID column in x. Default: "id".

component_labels

A character vector of labels to assign to components. Default: greekletters[[1]]

...

Additional arguments passed on to similar_soils(), such as threshold_single or threshold_all.

Value

A data.frame containing the original data along with a new 'compname' column that assigns a map unit component label to each group of similar soils.

Details

This function automates the iterative process of calling similar_soils() to build out an estimate of map unit composition.

As a special case, if threshold_single and threshold_all are set to 0, the function will use a more efficient method. It will directly group all soils that share the exact same combination of rated properties, as this is the logical outcome of a zero-distance threshold. The component labels are assigned based on the size of the groups, with the largest group being assigned the first label.

Examples

data("loamy", package = "SOILmilaR")

rate_taxpartsize <- function(x) {
  dplyr::case_match(x,
                    c("sandy-skeletal") ~ 1,
                    c("sandy") ~ 3,
                    c("loamy", "coarse-loamy", "coarse-silty") ~ 5,
                    c("fine-loamy", "fine-silty") ~ 7,
                    c("clayey", "fine") ~ 9,
                    c("very-fine") ~ 11,
                    c("loamy-skeletal", "clayey-skeletal") ~ 13,
                    "fragmental" ~ 15)
}

rate_depthclass <- function(x,
                            breaks = c(
                              `very shallow` = 25,
                              `shallow` = 50,
                              `moderately deep` = 100,
                              `deep` = 150,
                              `very deep` = 1e4
                            ),
                            ...) {
  res <- cut(x, c(0, breaks))
  factor(res, levels = levels(res), labels = names(breaks))
}

m <- list(taxpartsize = rate_taxpartsize, depth = rate_depthclass)

mapunit_composition <- design_mapunit(loamy, m)

mapunit_composition[order(mapunit_composition$component), ]
#>    id    taxpartsize     depth pscs_clay pscs_frags component
#> 1  A1     fine-loamy  68.07141  34.26617 13.3430897     Alpha
#> 2  B1     fine-loamy 125.65509  25.70668 10.3920511     Alpha
#> 3  C1     fine-loamy  82.03235  29.51870  9.6076022     Alpha
#> 4  D1     fine-loamy 136.54700  27.73477 33.8624746     Alpha
#> 11 A2     fine-loamy 145.74779  20.42760  0.6874675     Alpha
#> 12 B2     fine-loamy 138.76439  25.04729  6.6330011     Alpha
#> 13 C2     fine-loamy 114.43111  25.03331 11.9838727     Alpha
#> 14 D2     fine-loamy 126.47875  24.27037 12.9255822     Alpha
#> 21 A3     fine-loamy 111.48825  30.82608  3.6542921     Alpha
#> 23 C3     fine-loamy  79.15651  30.07310  6.2647017     Alpha
#> 24 D3     fine-loamy  66.55412  18.01062 28.9167000     Alpha
#> 8  H1 loamy-skeletal 137.62819  14.16824 49.8535505      Beta
#> 9  I1 loamy-skeletal  98.41503  15.31168 82.2289934      Beta
#> 10 J1 loamy-skeletal  87.51069  17.81801 38.6778412      Beta
#> 18 H2 loamy-skeletal  59.88691  15.86385 53.8326966      Beta
#> 20 J2 loamy-skeletal  61.63697  17.43131 44.3615694      Beta
#> 28 H3 loamy-skeletal 128.42479  16.45108 57.3262779      Beta
#> 30 J3 loamy-skeletal  85.58064  14.44454 39.3763163      Beta
#> 6  F1          loamy  40.23900  17.59930 27.0047312     Delta
#> 15 E2          loamy  37.83057  14.60978 23.4627516     Delta
#> 22 B3          loamy  45.90668  28.69676 10.0208338     Delta
#> 19 I2     fragmental  71.59082  15.06389 97.3761340   Epsilon
#> 29 I3     fragmental 126.34937  15.40719 97.1617265   Epsilon
#> 5  E1   coarse-loamy 143.15374  14.41170 25.7369392     Gamma
#> 7  G1   coarse-loamy  95.73213  14.98435 23.0575846     Gamma
#> 16 F2   coarse-loamy  89.94654  14.55522 14.9567534     Gamma
#> 17 G2   coarse-loamy 122.22285  14.93214 13.0607596     Gamma
#> 25 E3   coarse-loamy 128.68360  15.90127 12.4687515     Gamma
#> 26 F3   coarse-loamy  86.57938  14.88048 20.4374258     Gamma
#> 27 G3   coarse-loamy 128.15740  15.51927 33.6389675     Gamma

sort(prop.table(table(mapunit_composition$component)), decreasing = TRUE)
#> 
#>      Alpha       Beta      Gamma      Delta    Epsilon 
#> 0.36666667 0.23333333 0.23333333 0.10000000 0.06666667