Similar Soil Contrasts

A method for applying standardized, customizable "similar soils" rules to site-level data derived from various sources.

Usage

similar_soils(
  x,
  mapping,
  condition = NULL,
  idname = "id",
  thresh = NULL,
  thresh_single = 2,
  thresh_all = 3,
  absolute = TRUE,
  verbose = TRUE
)

Arguments

x

A data.frame or a SoilProfileCollection

mapping

A named list of functions. List element names refer to site-level data columns of x. Each specified function provides a conversion of the data element in x to a value used in the "similar soils" calculation.

condition

integer or character. Default: NULL the value is calculated internally based on the dominant condition of intersection of mapping results in x. You may specify an integer row ID in x for specific similar soil contrasts, or you can specify a character dominant condition equivalent to the name assigned by interaction() e.g. "4.3" for a two rating mapping result where the first rating has value 4 and the second has value 3

idname

ID column name, default "id"

thresh

Deprecated. If used passed to thresh_all.

thresh_single

Sum of differences relative to dominant condition in x. Default 1. See details for discussion on the default calculation for similarity.

thresh_all

Sum of differences relative to dominant condition in x. Default 1. See details for discussion on the default calculation for similarity.

absolute

logical. Report absolute difference? Default: TRUE. Absolute difference is always used for comparison against thresh.

verbose

Default: TRUE message about selected condition

Value

A data.frame containing inputs and three new columns: similar_single (maximum difference in any one property, relative to condition), similar_dist (cumulative sum of differences relative to condition), similar (logical; soil is similar to condition)

Details

The sum of differences across conditions (specified by the intersection of output of the functions in mapping) is used as the "distance" of a soil relative to a dominant (or otherwise specified) condition. A threshold value is used to decide which are "similar" and which are not. The functions in mapping can be customized to use alternate thresholds.

References

Norfleet, M.L. and Eppinette, R.T. (1993), A Mathematical Model for Determining Similar and Contrasting Inclusions for Map Unit Descriptons. Soil Survey Horizons, 34: 4-5. https://doi.org/10.2136/sh1993.1.0004

Examples